# openapi/tktsync.v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TktSync API",
    "version": "1.0.0",
    "description": "Authoritative TktSync API contract for implemented M3, M4, and M5 Admin and Partner surfaces."
  },
  "paths": {
...
    "/api/v1/realtime/stream": {
      "get": {
        "operationId": "realtimeStream",
        "security": [
          {
            "ScannerBearer": []
          }
        ],
        "parameters": [
          {
            "name": "event_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^evt_"
            }
          },
          {
            "name": "audience",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["admin", "scanner"]
            }
          },
          {
            "$ref": "#/components/parameters/XRequestID"
          }
        ],
        "responses": {
          "200": {
            "description": "Best-effort post-commit invalidation stream; clients must refetch authoritative state.",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/XRequestID"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/CacheControlNoStore"
              }
            },
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorResponse"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "PartnerBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "ScannerBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "SelectionBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Opaque, time-bounded BuyerSelectionSession capability."
      }
    },
    "parameters": {
      "XRequestID": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "Optional client request/correlation UUID.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "responses": {
      "ErrorResponse": {
        "description": "Machine-readable error response.",
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/XRequestID"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": ["code", "message", "request_id", "details"],
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "request_id": {
                "type": "string",
                "format": "uuid"
              },
              "details": {
                "oneOf": [
                  {
                    "type": "object",
                    "additionalProperties": true
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          }
        }
      },
...
    },
    "headers": {
      "XRequestID": {
        "description": "Request/correlation identifier generated or normalized by TktSync.",
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "CacheControlNoStore": {
        "description": "Reservation responses containing protected state must not be cached.",
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      }
    }
  }
}